# Optimizer: Run BetweenComposition and ChunkPruning rules bottom-up