--- a/src/lib/optimizer/strategy/abstract_rule.hpp
+++ b/src/lib/optimizer/strategy/abstract_rule.hpp
@@ -47,7 +47,12 @@ class AbstractRule {
  protected:
   /**
    * This function applies the concrete rule to the given plan, but not to its subquery plans.
-   * To traverse LQPs, use the visit_lqp function. Do not call this function recursively.
+   *
+   * DO NOT CALL THIS FUNCTION RECURSIVELY!
+   *  The reason for this can be found in diamond LQPs: When using "trivial" recursion, we would go down both on the
+   *  left and the right side of the diamond. On both sides, we would reach the bottom of the diamond. From there, we
+   *  would look at each node twice. visit_lqp prevents this by tracking which nodes have already been visited and
+   *  avoiding visiting a node twice.
    */
   virtual void _apply_to_plan_without_subqueries(const std::shared_ptr<AbstractLQPNode>& lqp_root) const = 0;
 };