--- a/src/lib/optimizer/strategy/between_composition_rule.cpp
+++ b/src/lib/optimizer/strategy/between_composition_rule.cpp
@@ -19,7 +19,7 @@ using namespace opossum::expression_functional;  // NOLINT
 namespace {
 using namespace opossum;  // NOLINT
 
-static PredicateCondition get_between_predicate_condition(bool left_inclusive, bool right_inclusive) {
+PredicateCondition get_between_predicate_condition(bool left_inclusive, bool right_inclusive) {
   if (left_inclusive && right_inclusive) {
     return PredicateCondition::BetweenInclusive;
   } else if (left_inclusive && !right_inclusive) {