… src/lib/optimizer/strategy/chunk_pruning_rule.cpp
… src/lib/optimizer/strategy/chunk_pruning_rule.hpp
… src/lib/optimizer/strategy/expression_reduction_rule.cpp
… src/lib/optimizer/strategy/stored_table_column_alignment_rule.cpp
… src/lib/optimizer/strategy/subquery_to_join_rule.cpp
… src/lib/optimizer/strategy/subquery_to_join_rule.hpp
… src/lib/statistics/cardinality_estimator.cpp
… src/lib/statistics/statistics_objects/abstract_histogram.cpp
… src/test/lib/logical_query_plan/alias_node_test.cpp
… src/test/lib/logical_query_plan/lqp_utils_test.cpp
… src/test/lib/optimizer/optimizer_test.cpp
… src/test/lib/optimizer/strategy/between_composition_rule_test.cpp