- [x] Refactoring
  - [x] BetweenCompositionRule (rudimentary)
  - [x] ChunkPruningRule @j-tr @julianmenzler 
    - [x] ~~Try to make `_compute_exclude_list` "read-only" (Currently, it modifies the table statistics of StoredTableNodes)~~
    - [x] Add Join-related tests @j-tr 
    - [x] Create generic utility functions and move them into lqp_utils.cpp ?
  - [x] BetweenCompositionRule (make it nice and consistent with ChunkPruningRule)
  - [x] SubqueryToJoinRule
  - [x] OptimizerTest
- [x] Remove _apply_to_inputs_without_subqueries 
- [x] Wait for #2267
- [x] Address Feedback
- [x] Fix CI
  - [x] Investigate issues with JoinOrder Benchmark
  - [x] Fix CardinalityEstimator
- [x] Run benchmark_all.sh 
- [x] Fix headlines in BM results
- [x] ChunkPruningRule Assert "did not expect to see node twice": Fix tests
- [x] Write PR description
- [x] Re-run benchmark_all.sh 

---

This PR removes the legacy function `_apply_to_plan_inputs_without_subqueries`, which was based on recursion and prone to errors when used. (Fixes #2276)

Due to the removal of the above function from `AbstractRule`, the following optimizer rules needed refactoring.

### SubqueryToJoinRule
The SubqueryToJoinRule did not change much. To traverse the LQP from top to bottom, it now uses the `visit_lqp` function and tracks all visited nodes to prevent any revisitation.

### BetweenCompositionRule & ChunkPruningRule
The BetweenCompositionRule and the ChunkPruningRule received a new traversal approach. They now traverse LQPs in a bottom-up manner using `visit_lqp_upwards`. For the ChunkPruningRule, this means that in some queries, more chunks can be pruned than before.

--- 

## Benchmark Results

**System**
<details>
<summary>pella - click to expand</summary>

| property | value |
| -- | -- |
| Hostname | pella |
| CPU | Intel(R) Xeon(R) CPU E7- 8870  @ 2.40GHz |
| Memory | 1.3TB |
| numactl | nodebind: 2  |
| numactl | membind: 2  |
</details>

**Commit Info and Build Time**
| commit | date | message | build time |
| -- | -- | -- | -- |
| 78a4342ae | 10.03.2021 16:02 | sql-parser: Fix unary minus (#2324) | real 848.22 user 5188.48 sys 243.72|
| 772d30f54 | 11.03.2021 01:07 | Fix SubqueryToJoinRule, finally | real 827.56 user 4913.00 sys 223.04|


**hyriseBenchmarkTPCH - single-threaded, SF 10.0**
<details>
<summary>
Sum of avg. item runtimes: -0%
 || 
Geometric mean of throughput changes: +0%
</summary>

<details>
<summary>Configuration Overview - click to expand</summary>

```diff
 +Configuration Overview----+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
 | Parameter                | /home/Julian.Menzler/repo_hyrise2/cmake_release/benchmark_all_results/hyriseBenchmarkTPCH_78a4342ae76d3bcd19442557ab0105cf12bef161_st.json | /home/Julian.Menzler/repo_hyrise2/cmake_release/benchmark_all_results/hyriseBenchmarkTPCH_772d30f5441af2266ae27d5ba4c79ddd3953ad68_st.json |
 +--------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
 |  GIT-HASH                | 78a4342ae76d3bcd19442557ab0105cf12bef161-dirty                                                                                             | 772d30f5441af2266ae27d5ba4c79ddd3953ad68-dirty                                                                                             |
 |  benchmark_mode          | Ordered                                                                                                                                    | Ordered                                                                                                                                    |
 |  build_type              | release                                                                                                                                    | release                                                                                                                                    |
 |  chunk_size              | 65535                                                                                                                                      | 65535                                                                                                                                      |
 |  clients                 | 1                                                                                                                                          | 1                                                                                                                                          |
 |  clustering              | None                                                                                                                                       | None                                                                                                                                       |
 |  compiler                | gcc 9.2                                                                                                                                    | gcc 9.2                                                                                                                                    |
 |  cores                   | 0                                                                                                                                          | 0                                                                                                                                          |
 |  date                    | 2021-03-11 01:24:42                                                                                                                        | 2021-03-11 09:42:25                                                                                                                        |
 |  encoding                | {'default': {'encoding': 'Dictionary'}}                                                                                                    | {'default': {'encoding': 'Dictionary'}}                                                                                                    |
 |  indexes                 | False                                                                                                                                      | False                                                                                                                                      |
 |  max_duration            | 60000000000                                                                                                                                | 60000000000                                                                                                                                |
 |  max_runs                | -1                                                                                                                                         | -1                                                                                                                                         |
 |  scale_factor            | 10.0                                                                                                                                       | 10.0                                                                                                                                       |
 |  time_unit               | ns                                                                                                                                         | ns                                                                                                                                         |
 |  use_prepared_statements | False                                                                                                                                      | False                                                                                                                                      |
 |  using_scheduler         | False                                                                                                                                      | False                                                                                                                                      |
 |  verify                  | False                                                                                                                                      | False                                                                                                                                      |
 |  warmup_duration         | 0                                                                                                                                          | 0                                                                                                                                          |
 +--------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
```
</details>

```diff
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Item     || Latency (ms/iter)  | Change || Throughput (iter/s) | Change | p-value |
 |          ||      old |     new |        ||      old |      new |        |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | TPC-H 01 ||  11291.5 | 11081.4 |   -2%  ||     0.09 |     0.09 |   +2%  |       ˅ |
 | TPC-H 02 ||     84.0 |    82.6 |   -2%  ||    11.90 |    12.10 |   +2%  |  0.0052 |
 | TPC-H 03 ||   4280.4 |  4261.5 |   -0%  ||     0.23 |     0.23 |   +0%  |  0.5838 |
 | TPC-H 04 ||   3085.8 |  3123.0 |   +1%  ||     0.32 |     0.32 |   -1%  |  0.0908 |
 | TPC-H 05 ||   6851.7 |  6826.1 |   -0%  ||     0.15 |     0.15 |   +0%  |       ˅ |
 | TPC-H 06 ||    373.5 |   374.8 |   +0%  ||     2.68 |     2.67 |   -0%  |  0.5949 |
 | TPC-H 07 ||   1869.7 |  1894.0 |   +1%  ||     0.53 |     0.53 |   -1%  |  0.0338 |
 | TPC-H 08 ||   1539.2 |  1542.9 |   +0%  ||     0.65 |     0.65 |   -0%  |  0.2679 |
 | TPC-H 09 ||  11751.0 | 11754.2 |   +0%  ||     0.09 |     0.09 |   -0%  |       ˅ |
 | TPC-H 10 ||   4792.5 |  4812.6 |   +0%  ||     0.21 |     0.21 |   -0%  |  0.6443 |
 | TPC-H 11 ||    200.0 |   197.4 |   -1%  ||     5.00 |     5.07 |   +1%  |  0.0000 |
 | TPC-H 12 ||   2057.2 |  2053.1 |   -0%  ||     0.49 |     0.49 |   +0%  |  0.5140 |
 | TPC-H 13 ||  11013.7 | 10948.1 |   -1%  ||     0.09 |     0.09 |   +1%  |       ˅ |
 | TPC-H 14 ||    959.4 |   953.9 |   -1%  ||     1.04 |     1.05 |   +1%  |  0.1538 |
 | TPC-H 15 ||    499.4 |   495.9 |   -1%  ||     2.00 |     2.02 |   +1%  |  0.0002 |
 | TPC-H 16 ||   1341.2 |  1315.1 |   -2%  ||     0.75 |     0.76 |   +2%  |  0.0000 |
 | TPC-H 17 ||    572.7 |   569.5 |   -1%  ||     1.75 |     1.76 |   +1%  |  0.0096 |
 | TPC-H 18 ||   5059.4 |  5065.2 |   +0%  ||     0.20 |     0.20 |   -0%  |  0.8807 |
 | TPC-H 19 ||    676.6 |   668.5 |   -1%  ||     1.48 |     1.50 |   +1%  |  0.0036 |
 | TPC-H 20 ||    915.3 |   916.4 |   +0%  ||     1.09 |     1.09 |   -0%  |  0.9239 |
 | TPC-H 21 ||  10413.7 | 10332.2 |   -1%  ||     0.10 |     0.10 |   +1%  |       ˅ |
 | TPC-H 22 ||    951.8 |   946.3 |   -1%  ||     1.05 |     1.06 |   +1%  |  0.0009 |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Sum      ||  80579.8 | 80214.8 |   -0%  ||          |          |        |         |
 | Geomean  ||          |         |        ||          |          |   +0%  |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 |          || ˅ Insufficient number of runs for p-value calculation                 |
 +----------++----------+---------+--------++----------+----------+--------+---------+
```
</details>


**hyriseBenchmarkTPCH - single-threaded, SF 0.01**
<details>
<summary>
Sum of avg. item runtimes: -0%
 || 
Geometric mean of throughput changes: -0%
</summary>

<details>
<summary>Configuration Overview - click to expand</summary>

```diff
 +Configuration Overview----+------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------------------------------------------------+
 | Parameter                | /home/Julian.Menzler/repo_hyrise2/cmake_release/benchmark_all_results/hyriseBenchmarkTPCH_78a4342ae76d3bcd19442557ab0105cf12bef161_st_s01.json | /home/Julian.Menzler/repo_hyrise2/cmake_release/benchmark_all_results/hyriseBenchmarkTPCH_772d30f5441af2266ae27d5ba4c79ddd3953ad68_st_s01.json |
 +--------------------------+------------------------------------------------------------------------------------------------------------------------------------------------+------------------------------------------------------------------------------------------------------------------------------------------------+
 |  GIT-HASH                | 78a4342ae76d3bcd19442557ab0105cf12bef161-dirty                                                                                                 | 772d30f5441af2266ae27d5ba4c79ddd3953ad68-dirty                                                                                                 |
 |  benchmark_mode          | Ordered                                                                                                                                        | Ordered                                                                                                                                        |
 |  build_type              | release                                                                                                                                        | release                                                                                                                                        |
 |  chunk_size              | 65535                                                                                                                                          | 65535                                                                                                                                          |
 |  clients                 | 1                                                                                                                                              | 1                                                                                                                                              |
 |  clustering              | None                                                                                                                                           | None                                                                                                                                           |
 |  compiler                | gcc 9.2                                                                                                                                        | gcc 9.2                                                                                                                                        |
 |  cores                   | 0                                                                                                                                              | 0                                                                                                                                              |
 |  date                    | 2021-03-11 01:50:26                                                                                                                            | 2021-03-11 10:08:03                                                                                                                            |
 |  encoding                | {'default': {'encoding': 'Dictionary'}}                                                                                                        | {'default': {'encoding': 'Dictionary'}}                                                                                                        |
 |  indexes                 | False                                                                                                                                          | False                                                                                                                                          |
 |  max_duration            | 60000000000                                                                                                                                    | 60000000000                                                                                                                                    |
 |  max_runs                | -1                                                                                                                                             | -1                                                                                                                                             |
 |  scale_factor            | 0.009999999776482582                                                                                                                           | 0.009999999776482582                                                                                                                           |
 |  time_unit               | ns                                                                                                                                             | ns                                                                                                                                             |
 |  use_prepared_statements | False                                                                                                                                          | False                                                                                                                                          |
 |  using_scheduler         | False                                                                                                                                          | False                                                                                                                                          |
 |  verify                  | False                                                                                                                                          | False                                                                                                                                          |
 |  warmup_duration         | 0                                                                                                                                              | 0                                                                                                                                              |
…