@julianmenzler 2021-01-19
All tests pass. But, the `hyriseJoinOrderBenchmark` [throws the following](https://hyrise-ci.epic-hpi.de/blue/organizations/jenkins/hyrise%2Fhyrise/detail/PR-2284/21/pipeline/97):

```
- Starting Benchmark...
terminate called after throwing an instance of 'std::logic_error'
  what():  src/lib/statistics/cardinality_estimator.cpp:986 Semi join should not increase cardinality
Traceback (most recent call last):
...
```

Not sure, whats wrong here yet. My assumption: It is because of the refactored BetweenCompositionRule. 

To make sure: @j-tr 
Have you already pulled the ChunkPruningRule changes into your branch? If so, does hyriseJoinOrderBenchmark pass?