@mrks 2021-01-19
That assertion is somewhat flaky. The underlying principle is true, but we have seen it popping up in unexpected places before. You could print the estimated cardinalities for the master and your branch to see if they actually diverge significantly.