@julianmenzler 2021-02-24
The CI issue
```
src/lib/statistics/cardinality_estimator.cpp:987 Semi join should not increase cardinality\r\n'
```

has been solved. However, some LQPs still look weird because they contain negative row counts. See, for example, Q23c in the JoinOrderBenchmark:

<img width="916" alt="image" src="https://user-images.githubusercontent.com/9380917/109029911-9bd41580-76c3-11eb-8bdc-8ff08712f2f0.png">