@j-tr 2021-02-24
... not only negative but also negative non integer row counts. So far, i tracked down that the BetweenCompositionRule introduces float row counts and the ChunkPruningRule negates some of them.