@mrks 2021-02-24
Non-integer is not a problem, but an expected side effect of the statistics' uncertainty.