@mrks 2021-02-24
Negative row counts, however, should not occur. Maybe the GenericHistogramBuilder should assert that numbers are not negative, not NaN etc