@j-tr 2021-02-24
so 0.1 row(s) is an expected outcome that we should not worry about?