@mrks 2021-02-24
Yes. Imagine a table with two columns, 10 rows, and all entries being distinct. Filtering by the first column gives us a selectivity of 10%. With this, we expect 1 row (10*0.1) as a result. If we then do the same for the second column, we end up at 0.1.

Does this make sense? Debatable, but this is the current approach.

If you can think of a good position for a small comment, feel free to add one.