@mrks 2021-02-24
> thanks for explaining. i guess thats also what caused the negative row counts. when the row count is based on an estimation and the new row count after chunk pruning is calculated as (estimated row count - actually pruned rows), we can get a negative row count.

Mmh, I see. In that case, we should return a cardinality of 0.