@Bouncner 2021-03-03
Github müffelt. Not sure if the following comments are properly shown: https://github.com/hyrise/hyrise/commit/b7bf3dff91cc62e8fb6a0d1b6ab4b2c3348b0e54#comments