@julianmenzler 2021-03-08
Not sure what causes the performance drop of TPC-H Q13 & Q14 in single-threaded mode.