@julianmenzler 2021-03-09
**Master**
![tpch_master_optimizer_breakdown](https://user-images.githubusercontent.com/9380917/110469351-d2555b80-80d9-11eb-8ca6-755da60f331c.png)

**This PR**
![tpch_2284_optimizer_breakdown](https://user-images.githubusercontent.com/9380917/110469344-d1242e80-80d9-11eb-9d99-e618e8e906fd.png)

The ChunkPruningRule takes more time than before. However, according to the following chart, the optimization time is negligible:
![tpch_2284_breakdown](https://user-images.githubusercontent.com/9380917/110470007-9ec70100-80da-11eb-83ea-f034d4405bd0.png)