@julianmenzler 2021-03-10
The Assert

```cpp
Assert(!visited_nodes.contains(current_node), "Did not expect to see the same node twice.");
```

 fails in both rules, ChunkPruningRule & SubqueryToJoinRule.

Looking into it...