@julianmenzler 2021-03-11
> Just quick requests for documentation. Do you have benchmark results that include the current master?

In progress.