@julianmenzler 2021-03-11
Benchmark results look really good! No performance drop for TPC-H Q13 & Q14 anymore :)